Refactore for firebase
