+++ client/src/components/common/CommentList.jsx
@@ -1,6 +1,5 @@
-// components/comments/CommentList.jsx
 import { useEffect, useState } from "react";
-import { db } from "../../services/firebase";
+import { db } from "../../services/firebase/firebase";
 import { collection, query, orderBy, onSnapshot } from "firebase/firestore";
 
 const CommentList = ({ slug }) => {
